Make utils: ForwardAdaptor
